split graphql soup into modules
